comparing char types with EOF can be dangerous
An unsigned char can never equal EOF. If comparing EOF against a
signed char, One could get a premature termination because some data
character happened to be all ones.

Standard character functions like getc() return an int and not a char
because it must be capable of returning 257 different values (256
different characters plus EOF, assuming an 8-bit character).  Once
this value is assigned to a char only 256 values are then possible --
a clear loss of information.